# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_financeos_api_internal_usecase.CreateAccountRequest:
    properties:
      balance:
        type: number
      color:
        type: string
      credit_limit:
        type: number
      icon:
        type: string
      institution:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      type:
        enum:
        - checking
        - savings
        - credit_card
        - investment
        - wallet
        - other
        type: string
    required:
    - name
    - type
    type: object
...
  internal_handler.AccountListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.Account'
        type: array
    type: object
  internal_handler.AccountResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.Account'
    type: object
...
  internal_handler.AccountSummaryResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.AccountSummaryData'
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorDetail'
    type: object
...
  internal_handler.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.MessageData'
    type: object
...
host: localhost:8000
info:
  contact:
    name: FinanceOS Support
    url: https://github.com/financeos
  description: API completa para gestão financeira pessoal — contas, transações, orçamentos,
    investimentos, metas e IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinanceOS API
  version: 1.0.0
paths:
  /accounts:
    get:
      description: Retorna todas as contas ativas do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar contas
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Cria uma nova conta bancária, carteira ou cartão
      parameters:
      - description: Dados da conta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar conta
      tags:
      - Accounts
  /accounts/{id}:
    delete:
      parameters:
      - description: UUID da conta
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Desativar conta
      tags:
      - Accounts
    get:
      parameters:
      - description: UUID da conta
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar conta por ID
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID da conta
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar conta
      tags:
      - Accounts
  /accounts/summary:
    get:
      description: Retorna saldo total e saldo por conta
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AccountSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resumo de saldo
      tags:
      - Accounts
  /ai/chat:
    post:
      consumes:
      - application/json
      description: Conversa com o assistente financeiro alimentado por Claude
      parameters:
      - description: Mensagem (máx. 2000 chars)
        in: body
        name: body
        required: true
        schema:
          properties:
            message:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AIChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assistente financeiro
      tags:
      - AI
  /ai/portfolio-analysis:
    get:
      description: Insights sobre diversificação e performance (plano Pro)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AIForecastResponse'
        "402":
          description: Plano insuficiente
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Análise de portfólio com IA
      tags:
      - AI
  /ai/spending-forecast:
    get:
      description: Análise preditiva de gastos com IA (plano Pro)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AIForecastResponse'
        "402":
          description: Plano insuficiente
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Previsão de gastos
      tags:
      - AI
  /assets/search:
    get:
      parameters:
      - description: 'Ticker ou nome (ex: PETR4)'
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AssetSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar ativos por ticker ou nome
      tags:
      - Investments
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Envia e-mail com link de redefinição (sempre retorna 200 por segurança)
      parameters:
      - description: E-mail
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Solicitar redefinição de senha
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica o usuário e retorna tokens JWT
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AuthDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Credenciais inválidas
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoga os tokens do usuário e encerra a sessão
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Troca um refresh token expirado por um novo par de tokens
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AuthDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Token inválido ou revogado
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Renovar tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cria uma conta e retorna tokens JWT + dados do usuário
      parameters:
      - description: Dados de registro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.AuthDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: E-mail já cadastrado
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Registrar novo usuário
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Define uma nova senha usando o token recebido por e-mail
      parameters:
      - description: Token e nova senha (mín. 8 caracteres)
        in: body
        name: body
        required: true
        schema:
          properties:
            password:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "400":
          description: Token inválido ou senha fraca
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Redefinir senha
      tags:
      - Auth
  /budgets:
    get:
      parameters:
      - description: 'Mês (1-12, default: atual)'
        in: query
        name: month
        type: integer
      - description: 'Ano (default: atual)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.BudgetListResponse'
      security:
      - BearerAuth: []
      summary: Listar orçamentos
      tags:
      - Budgets
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do orçamento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar orçamento
      tags:
      - Budgets
  /budgets/{id}:
    delete:
      parameters:
      - description: UUID do orçamento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover orçamento
      tags:
      - Budgets
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID do orçamento
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.UpdateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.BudgetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar orçamento
      tags:
      - Budgets
  /budgets/progress:
    get:
      description: Retorna planejado vs real por categoria
      parameters:
      - description: 'Mês (1-12, default: atual)'
        in: query
        name: month
        type: integer
      - description: 'Ano (default: atual)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.BudgetProgressResponse'
      security:
      - BearerAuth: []
      summary: Progresso dos orçamentos
      tags:
      - Budgets
  /categories:
    get:
      description: Retorna categorias do sistema e categorias personalizadas do usuário
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CategoryListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar categorias
      tags:
      - Categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da categoria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar categoria
      tags:
      - Categories
  /categories/{id}:
    delete:
      parameters:
      - description: UUID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover categoria
      tags:
      - Categories
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar categoria
      tags:
      - Categories
  /custom-assets:
    get:
      description: Retorna imóveis, veículos e outros ativos cadastrados manualmente
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CustomAssetListResponse'
      security:
      - BearerAuth: []
      summary: Listar ativos personalizados
      tags:
      - Investments
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do ativo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.CreateCustomAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.CustomAssetResponse'
      security:
      - BearerAuth: []
      summary: Criar ativo personalizado
      tags:
      - Investments
  /custom-assets/{id}:
    delete:
      parameters:
      - description: UUID do ativo
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
      security:
      - BearerAuth: []
      summary: Remover ativo personalizado
      tags:
      - Investments
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID do ativo
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.UpdateCustomAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CustomAssetResponse'
      security:
      - BearerAuth: []
      summary: Atualizar ativo personalizado
      tags:
      - Investments
  /dashboard/cashflow:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CashflowResponse'
      security:
      - BearerAuth: []
      summary: Cashflow dos últimos 12 meses
      tags:
      - Dashboard
  /dashboard/overview:
    get:
      description: Retorna saldo líquido, receitas, despesas e top categorias
      parameters:
      - description: 'Mês (1-12, default: atual)'
        in: query
        name: month
        type: integer
      - description: 'Ano (default: atual)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.DashboardOverviewResponse'
      security:
      - BearerAuth: []
      summary: Visão geral do mês
      tags:
      - Dashboard
  /family:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.FamilyGroupResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter grupo familiar
      tags:
      - Family
    post:
      consumes:
      - application/json
      parameters:
      - description: Nome do grupo
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.FamilyGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar grupo familiar
      tags:
      - Family
  /family/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.FamilyDashboardResponse'
      security:
      - BearerAuth: []
      summary: Dashboard do grupo familiar
      tags:
      - Family
  /family/invite:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.InviteCodeResponse'
      security:
      - BearerAuth: []
      summary: Gerar código de convite
      tags:
      - Family
  /family/join:
    post:
      consumes:
      - application/json
      parameters:
      - description: Código de convite
        in: body
        name: body
        required: true
        schema:
          properties:
            invite_code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.FamilyGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Entrar em grupo familiar via convite
      tags:
      - Family
  /family/members/{id}:
    delete:
      parameters:
      - description: UUID do membro
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Remover membro do grupo
      tags:
      - Family
  /goals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.GoalListResponse'
      security:
      - BearerAuth: []
      summary: Listar metas
      tags:
      - Goals
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da meta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.CreateGoalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.GoalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar meta
      tags:
      - Goals
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"